+++ src/main/java/core/basesyntax/model/IsoscelesTrapezoid.java
@@ -1,11 +1,11 @@
 package core.basesyntax.model;
 
 public class IsoscelesTrapezoid extends Figure {
-    private double firstBase;
-    private double secondBase;
-    private double height;
+    private int firstBase;
+    private int secondBase;
+    private int height;
 
-    public IsoscelesTrapezoid(String color, double firstBase, double secondBase, double height) {
+    public IsoscelesTrapezoid(String color, int firstBase, int secondBase, int height) {
         super(color);
         this.firstBase = firstBase;
         this.secondBase = secondBase;
@@ -15,15 +15,15 @@ public class IsoscelesTrapezoid extends Figure {
     @Override
     public void draw() {
         System.out.println("Figure: isosceles trapezoid, area: " + getArea()
-                + " sq.units, firstBase: " + (int)firstBase + " units,"
-                + " secondBase: " + (int)secondBase + " units,"
-                + " height: " + (int)height + " units,"
+                + " sq.units, firstBase: " + firstBase + " units,"
+                + " secondBase: " + secondBase + " units,"
+                + " height: " + height + " units,"
                 + " color: " + getColor());
     }
 
     @Override
     public double getArea() {
-        return ((firstBase + secondBase) / 2) * height;
+        return ((firstBase + secondBase) / 2.0) * height;
 
     }
 }
