+++ src/main/java/core/basesyntax/model/Square.java
@@ -1,9 +1,9 @@
 package core.basesyntax.model;
 
 public class Square extends Figure {
-    private double side;
+    private int side;
 
-    public Square(String color, double side) {
+    public Square(String color, int side) {
         super(color);
         this.side = side;
     }
@@ -11,7 +11,7 @@ public class Square extends Figure {
     @Override
     public void draw() {
         System.out.println("Figure: square, area: " + getArea()
-                + " sq.units, side: " + (int)side + " units,"
+                + " sq.units, side: " + side + " units,"
                 + " color: " + getColor());
     }
